// frontend/src/lib/api/models/teamContext.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * WinRate AI API
 * API de predictions de paris sportifs sur le football europeen
 * OpenAPI spec version: 0.1.0
 */

/**
 * Team context information.
 */
export interface TeamContext {
  team_name: string;
  recent_news?: string[];
  injuries?: string[];
  /**
   * @minimum -1
   * @maximum 1
   */
  sentiment_score?: number;
  sentiment_label?: string;
}